add support for --inputFile options

+++ RatePlots/trigger_plots.py
@@ -70,11 +70,12 @@ parser.add_argument('--vsRun', action='store_const', const=True, default=False, 
 parser.add_argument('--vsPU', action='store_const', const=True, default=False, help='Make plots vs pileup')
 parser.add_argument('--vsIntLumi', action='store_const', const=True, default=False, help='Make plots vs integrated luminosity. Selected by default if any --vs* flag is defined')
 parser.add_argument('--vsTime', action='store_const', const=True, default=False, help='Make plots vs days')
-parser.add_argument('--runMin', default=362104 , help='Run min. The minimum run possible run is 355678 (July 17, 2022)')
+parser.add_argument('--runMin', default=362104 , help='Select files with run>runMin. This option will be ignored when used with --inputFile. The minimum run possible run is 355678 (July 17, 2022)')
 parser.add_argument('--runMax', default=1000000 , help='Run max')
 parser.add_argument('--triggers', default="" , help='List of trigger used in the plots, separated by ",". If undefined, the triggerList defined in trigger_plots.py will be used. Example: --triggers HLT_IsoMu24_v,AlCa_PFJet40_v')
 parser.add_argument('--selections', default="" , help='List of selections used in the plots, separated by ",". If undefined, the triggerList defined in trigger_plots.py will be used.')
-parser.add_argument('--input', default="/afs/cern.ch/work/s/sdonato/public/OMS_ntuples/v2.0/" , help='Input folder containing the OMS ntuples')
+parser.add_argument('--input', default="" , help='Input folder containing the OMS ntuples. Cannot be used with --inputFile option. [Eg. /afs/cern.ch/work/s/sdonato/public/OMS_ntuples/v2.0/]')
+parser.add_argument('--inputFile', default="" , help='Input file containing the OMS ntuples. Cannot be used with --input option. [Eg. /afs/cern.ch/work/s/sdonato/public/OMS_ntuples/v2.0/goldejson_skim.root]')
 parser.add_argument('--output', default="plots/" , help='Folder of the output plots')
 parser.add_argument('--refLumi', default=2E34 , help='Reference rate used in the cross-section plots.')
 parser.add_argument('--lumisPerBin', default=1 , help='Number of lumisections that will be merged in the plots. Cannot work with --nbins')
@@ -86,10 +87,10 @@ parser.add_argument('--testing', action='store_const', const=True, default=False
 args = parser.parse_args()
 
 from tools import readOptions
-useRates, vses, triggers, folder, plotsFolder, removeOutliers, runMin, runMax, batch, testing, lumisPerBin, refLumi, selections, nbins = readOptions(parser.parse_args(), triggerDefault, selectionDefault)
+useRates, vses, triggers, inputFolder, inputFile, plotsFolder, removeOutliers, runMin, runMax, batch, testing, lumisPerBin, refLumi, selections, nbins = readOptions(parser.parse_args(), triggerDefault, selectionDefault)
 
 print("##### Options #####")
-print ("trigger_plots.py will produce %d x %d x %d x %d = %d plots in %s using OMS ntuples from %s,"%(len(useRates),len(vses),len(triggers),len(selections),len(useRates)*len(vses)*len(triggers)*len(selections), plotsFolder, folder))
+print ("trigger_plots.py will produce %d x %d x %d x %d = %d plots in %s using OMS ntuples from %s,"%(len(useRates),len(vses),len(triggers),len(selections),len(useRates)*len(vses)*len(triggers)*len(selections), plotsFolder, inputFolder if inputFolder else inputFile))
 if lumisPerBin>0: print("using %d lumisection per bin"%nbins)
 if nbins>0: print("using %d bins"%nbins)
 print(vses)
@@ -108,48 +109,49 @@ print("###################")
 
 ###################################################
 
-try:
-    files = os.listdir(folder)
-except:
-    print("#"*100)
-    print("Input folder %s not found. Please check your --input option."%folder)
-    print("You can download the OMS ntuple from OMS using:")
-    print("xrdcp --recursive root://eosuser.cern.ch//eos/user/s/sdonato/public/OMS_rates/v1.0/  .")
-    print("#"*100)
+if inputFile:
+    chain = ROOT.TChain("tree")
+    chain.AddFile(inputFile)
+else:
+    try:
+        files = os.listdir(inputFolder)
+    except:
+        print("#"*100)
+        print("Input folder %s not found. Please check your --input option."%inputFolder)
+        print("You can download the OMS ntuple from OMS using:")
+        print("xrdcp --recursive root://eosuser.cern.ch//eos/user/s/sdonato/public/OMS_rates/v1.0/  .")
+        print("#"*100)
 
 
-## Use only few files and few triggers for testing:
-if testing:
-    import shutil
-    selections = {"testing":selections["inclusive"]}
-    for f in selections:
-        shutil.rmtree(plotsFolder+"/"+f)
-    lumisPerBin = 1
-#    files = ["362655.root","357900.root"]
-#    triggers = ["HLT_DoubleMediumChargedIsoDisplacedPFTauHPS32_Trk1_eta2p1_v", "HLT_IsoMu24_v","AlCa_PFJet40_CPUOnly_v"]
-    runMin = 360449 # July 17. Different unit for rec lumi before this run.
-    runMax = 361500
-    vses = ["vsFill"]
-    useRates = [False]
-    folder = "/home/sdonato/CMS/OMS_plots/OMS_ntuples/"
-#    batch=False
+    ## Use only few files and few triggers for testing:
+    if testing:
+        import shutil
+        selections = {"testing":selections["inclusive"]}
+        for f in selections:
+            shutil.rmtree(plotsFolder+"/"+f)
+        lumisPerBin = 1
+    #    files = ["362655.root","357900.root"]
+    #    triggers = ["HLT_DoubleMediumChargedIsoDisplacedPFTauHPS32_Trk1_eta2p1_v", "HLT_IsoMu24_v","AlCa_PFJet40_CPUOnly_v"]
+        runMin = 360449 # July 17. Different unit for rec lumi before this run.
+        runMax = 361500
+        vses = ["vsFill"]
+        useRates = [False]
+        inputFolder = "/home/sdonato/CMS/OMS_plots/OMS_ntuples/"
+    #    batch=False
 
-runs = [int(f.split(".root")[0]) for f in files  if (f[0]=="3" and f[-5:]==".root")]
-#selection = "pileup_vs>53 && pileup_vs<57 && cms_ready && beams_stable && beam2_stable"
-#selection = "pileup_vs>54 && pileup_vs<56 && cms_ready && beams_stable && beam2_stable"
-#selection = "HLT_DoubleMediumChargedIsoDisplacedPFTauHPS32_Trk1_eta2p1_v>0"
+    runs = [int(f.split(".root")[0]) for f in files  if (f[0]=="3" and f[-5:]==".root")]
+    #selection = "pileup_vs>53 && pileup_vs<57 && cms_ready && beams_stable && beam2_stable"
+    #selection = "pileup_vs>54 && pileup_vs<56 && cms_ready && beams_stable && beam2_stable"
+    #selection = "HLT_DoubleMediumChargedIsoDisplacedPFTauHPS32_Trk1_eta2p1_v>0"
 
 
-## Load all not-empy files
-for run in sorted(runs):
-    fName = "%s/%d.root"%(folder,run)
-    if runMin>0 and run<runMin: continue
-    if runMax>0 and run>runMax: continue
-    if os.path.getsize(fName) > 1000 :
-        chain.AddFile(fName)
-
-#chain = ROOT.TChain("tree")
-#chain.AddFile("/afs/cern.ch/work/s/sdonato/public/OMS_ntuples/v2.0/goldejson_skim.root")
+    ## Load all not-empy files
+    for run in sorted(runs):
+        fName = "%s/%d.root"%(inputFolder,run)
+        if runMin>0 and run<runMin: continue
+        if runMax>0 and run>runMax: continue
+        if os.path.getsize(fName) > 1000 :
+            chain.AddFile(fName)
 
 ## Constants
 secInDay = 24.*60*60
@@ -157,9 +159,9 @@ LS_seconds = 2**18 / 11245.5
 LS_duration = LS_seconds/ secInDay #LS in days
 from datetime import datetime
 #offset = int(datetime(2023,1,1).timestamp()) - int(datetime(2022,8,31).timestamp())  #since Nov 1, 2022 instead of #since Jan 1, 2023
-try:
+try: ## python3 - reccomended
     offset = int(datetime(2023,1,1).timestamp()) - int(datetime(2022,8,31).timestamp())  #since Nov 1, 2022 instead of #since Jan 1, 2023
-except:
+except: ## python
     offset = 10630800
     print("Please use 'python3' instead of 'python',")
     print("using offset %d. This should be equal to int(datetime(2023,1,1).timestamp()) - int(datetime(2022,8,31).timestamp()) in python3 [from datetime import datetime] ")
@@ -316,7 +318,7 @@ for selFolder in selections:
     from style import getColor
     for i, trigger in enumerate(triggers[:]):
         print("Getting histo for ", trigger)
-        histos_vsTime[trigger] = getHisto("Alt$(%s,1)"%trigger, chain, timeVar, binning, selection) #Alt$(%s,1) ?
+        histos_vsTime[trigger] = getHisto("%s"%trigger, chain, timeVar, binning, selection) #Alt$(%s,1) ?
         histos_vsFill[trigger] = getHistoVsFillNumber(histos_vsTime[trigger], fillNumber_vsTime)
         histos_vsRun[trigger] = getHistoVsFillNumber(histos_vsTime[trigger], runNumber_vsTime)
         setStyle(histos_vsTime[trigger], getColor(i))
